+++ extensions/dashlane-vault/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Improvements] - 2024-04-23
 
 - Master password for Dashlane CLI can be saved in the extensions settings to lock your vault by default (Dashlane CLI version 6.2416.0 or newer required)
+- Biometrics for Dashlane CLI can now be used to access your vault (Dashlane CLI version 6.2416.0 or newer required)
 - Improved error handling
   - Errors are displayed to the user
   - Show error if user installed a buggy version (6.2415.0)